+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -2,9 +2,9 @@ package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.tobe.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
-import nextstep.mvc.controller.tobe.ManualHandlerAdapter;
+import nextstep.mvc.controller.tobe.handleradapter.AnnotationHandlerAdapter;
+import nextstep.mvc.controller.tobe.handlermapping.AnnotationHandlerMapping;
+import nextstep.mvc.controller.tobe.handleradapter.ManualHandlerAdapter;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
