+++ mvc/src/main/java/nextstep/mvc/controller/tobe/handleradapter/ManualHandlerAdapter.java
@@ -1,21 +1,24 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller.tobe.handleradapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
+import nextstep.mvc.controller.asis.Controller;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class AnnotationHandlerAdapter implements HandlerAdapter {
+public class ManualHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof HandlerExecution;
+        return handler instanceof Controller;
     }
 
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
                                final Object handler) throws Exception {
-        HandlerExecution handlerExecution = (HandlerExecution) handler;
-        return handlerExecution.handle(request, response);
+        Controller controller = (Controller) handler;
+        String viewName = controller.execute(request, response);
+        return new ModelAndView(new JspView(viewName));
     }
 }
