+++ mvc/src/test/java/nextstep/mvc/controller/tobe/ManualHandlerAdapterTest.java
@@ -7,6 +7,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import nextstep.mvc.controller.tobe.handleradapter.ManualHandlerAdapter;
+import nextstep.mvc.controller.tobe.handlermapping.HandlerExecution;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.Test;
